+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+<!-- 1.11.1 -->
+
+### Changes in v1.11.1
+
+- Improved the layout of `trading card manager`
+- Added an individual `sell` button for each card
+- Added a confirmation modal before listing trading cards
+- Added toasts for different actions, successes and errors in `trading card manager`
+- The `list selected cards` button now shows a loading state while cards a being listed
+
 <!-- 1.11.0 -->
 
 ### Changes in v1.11.0